requirements.txt
# Case 3 Automation Agent - Python Dependencies
# 根据 PROJECT_PLAN.md 和实施计划定义的依赖项

# ==================== 核心库 ====================
# Excel 处理
openpyxl>=3.0.0

# Word 处理
python-docx>=0.8.0

# 数据验证和模型
pydantic>=2.0.0

# ==================== AI 集成 ====================
# Anthropic Claude API
anthropic>=0.7.0

# ==================== 配置和工具 ====================
# YAML 配置文件处理
pyyaml>=6.0

# 重试机制（API 调用）
tenacity>=8.0.0

# ==================== 测试工具 ====================
# 测试框架
pytest>=7.0.0

# 测试覆盖率
pytest-cov>=4.0.0

# ==================== 开发工具（可选）====================
# 代码格式化
# black>=22.0.0

# 代码检查
# pylint>=2.15.0
# flake8>=5.0.0

# 类型检查
# mypy>=0.990
